requirements.txt
folium~=0.14.0
streamlit~=1.16.0
streamlit-folium
gpxpy~=1.5.0
geopandas~=0.12.2
tabulate~=0.9.0
pyyaml~=6.0
pandas~=1.5.2
# shapely~=2.0.0
# pillow~=9.4.0
requests~=2.28.1
plotly~=5.12.0
# owslib~=0.27.2